refactor!: moved Expression, ExpressionNode, utility functions to separate mods
...while maintaining their import path as `crate::expressions::{Expression, ExpressionNode}`.

The functions bool, var & vars are now part of the public API.
